feature: provide proper generics for `StorageInterface`
